Rename systemd.ts to dbus.ts
This is because it's more general, and just provides methods which act
over dbus.
